started working on the frontend aspect to display data from database and modifield some columns in the db

+++ app/Http/Controllers/Admin/PostController.php
@@ -4,11 +4,12 @@ namespace App\Http\Controllers\Admin;
 
 use App\Models\Post;
 use App\Models\Category;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\File;
 use App\Http\Requests\Admin\PostFormRequest;
-use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
@@ -34,6 +35,32 @@ class PostController extends Controller
         $post->name = $data['name'];
         $post->slug = Str::slug($data['slug']);
         $post->description = $data['description'];
+
+        if ($request->hasFile('poster')) {
+            $file = $request->file('poster');
+
+            $validExtension = ['jpg', 'png', 'jpeg'];
+            $maxFileSize = 2048;
+
+            if (!$file->isValid()) {
+                return redirect()->back()->with('error', 'Invalid File');
+            }
+
+            $extension = $file->getClientOriginalExtension();
+
+            if (!in_array(strtolower($extension), $validExtension)) {
+                return redirect()->back()->with('error', 'Invalid File extension');
+            }
+
+            if ($file->getSize() > $maxFileSize * 1024) {
+                return redirect()->back()->with('error', 'File Size is too large');
+            }
+
+            $filename = time() . '.' . $extension;
+            $file->move('admin/uploads/articles/', $filename);
+
+            $post->image = $file;
+        }
         $post->yt_iframe = $data['yt_iframe'];
         $post->meta_title = $data['meta_title'];
         $post->meta_description = $data['meta_description'];
@@ -66,6 +93,35 @@ class PostController extends Controller
         $post->name = $data['name'];
         $post->slug = Str::slug($data['slug']);
         $post->description = $data['description'];
+
+        if ($request->hasFile('poster')) {
+            $file = $request->file('poster');
+
+            $validExtension = ['jpg', 'png', 'jpeg'];
+            $maxFileSize = 2048;
+
+            if (!$file->isValid()) {
+                // Handle the case where the file is not valid (e.g., corrupted)
+                return redirect()->back()->with('error', 'Invalid file');
+            }
+
+            $extension = $file->getClientOriginalExtension();
+
+            if (!in_array(strtolower($extension), $validExtension)) {
+
+                return redirect()->back()->with('error', 'Invalid File Extension');
+            }
+
+            if ($file->getSize() > $maxFileSize * 1024) {
+
+                return redirect()->back()->with('error', 'File size is too large');
+            }
+
+            $filename = time() . '.' . $extension;
+            $file->move('admin/uploads/articles/', $filename);
+            $post->image->$file;
+        }
+
         $post->yt_iframe = $data['yt_iframe'];
         $post->meta_title = $data['meta_title'];
         $post->meta_description = $data['meta_description'];
@@ -82,8 +138,19 @@ class PostController extends Controller
     {
         $post = Post::find($post_id);
         if ($post) {
-            $post->delete();
-            return redirect()->route('admin.view_post')->with('error', 'Post deleted successful');
+            $destination = 'admin/uploads/articles/' . $post->image;
+            if (File::exist($destination)) {
+                if (File::delete($destination)) {
+                    $post->delete();
+                    return redirect()->route('admin.view_post')->with('error', 'Post deleted successful');
+                } else {
+                    return redirect()->route('admin.view_post')->with('error', 'Failed to delete the file');
+                }
+            } else {
+                return redirect()->route('admin.view_post')->with('error', 'File not found');
+            }
+        } else {
+            return redirect()->route('admin.view_post')->with('error', 'No Post Id found');
         }
     }
 }
